# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ruian-backend
  description: "\n    ## 锐安后端 API 文档\n\n    本 API 提供以下功能：\n\n    * **认证管理**: 用户登录和身份验证\n\
    \    * **文件上传**: 支持本地存储和阿里云 OSS 两种方式上传 Excel 文件\n    * **生产信息查询**: 查询和管理生产订单信息\n\
    \    * **员工工作量查询**: 查询和管理员工工作量记录\n    * **数据核验**: 验证生产信息和员工工作量数据的一致性\n\n    ##\
    \ 认证说明\n\n    大部分接口需要 Bearer Token 认证。请先调用 `/api/auth/login` 接口获取 token，\n   \
    \ 然后在请求头中添加：`Authorization: Bearer <your_token>`\n\n    ## 错误响应\n\n    所有错误响应都遵循统一格式：\n\
    \    ```json\n    {\n        \"error\": true,\n        \"message\": \"错误消息\",\n\
    \        \"detail\": {}\n    }\n    ```\n    "
  version: 0.1.0
paths:
  /api/auth/login:
    post:
      tags:
      - auth
      summary: 用户登录
      description: 使用用户名和密码登录，获取访问令牌（JWT Token）。
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Login Api Auth Login Post
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '422':
          description: 请求参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
  /api/auth/me:
    get:
      tags:
      - auth
      summary: 获取当前用户信息
      description: 获取当前已认证用户的基本信息。需要 Bearer Token 认证。
      operationId: get_current_user_info_api_auth_me_get
      security:
      - HTTPBearer: []
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 1
                username: admin
                created_at: '2025-01-01T00:00:00Z'
                updated_at: '2025-01-01T00:00:00Z'
        '401':
          description: 认证失败，Token 无效或已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/production:
    get:
      tags:
      - production
      summary: 查询生产信息列表
      description: 分页查询生产信息，支持按订单号、日期范围等条件过滤。需要 Bearer Token 认证。
      operationId: get_production_info_api_production_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 页码，从 1 开始
          default: 1
          title: Page
        description: 页码，从 1 开始
        example: 1
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 每页数量，最大 100
          default: 10
          title: Page Size
        description: 每页数量，最大 100
        example: 10
      - name: order_no
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 订单号（模糊匹配）
          title: Order No
        description: 订单号（模糊匹配）
        example: ORD20250101
      - name: start_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: 开始日期（基于 upload_date），格式：YYYY-MM-DDTHH:MM:SS
          title: Start Date
        description: 开始日期（基于 upload_date），格式：YYYY-MM-DDTHH:MM:SS
        example: '2025-01-01T00:00:00'
      - name: end_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: 结束日期（基于 upload_date），格式：YYYY-MM-DDTHH:MM:SS
          title: End Date
        description: 结束日期（基于 upload_date），格式：YYYY-MM-DDTHH:MM:SS
        example: '2025-01-31T23:59:59'
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionInfoListResponse'
              example:
                total: 100
                page: 1
                page_size: 10
                items:
                - id: 1
                  order_no: ORD20250101001
                  model: MODEL-A
                  brand_no: BRAND-001
                  quantity: 100
                  job_type: 组装
                  worklog_no: WL001
                  performance_factor: 1.2
                  upload_date: '2025-01-01T00:00:00Z'
                  created_at: '2025-01-01T00:00:00Z'
                  updated_at: '2025-01-01T00:00:00Z'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '422':
          description: 请求参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
  /api/production/{order_no}:
    get:
      tags:
      - production
      summary: 按订单号查询生产信息
      description: 根据订单号精确查询所有相关的生产信息记录。需要 Bearer Token 认证。
      operationId: get_production_info_by_order_no_api_production__order_no__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_no
        in: path
        required: true
        schema:
          type: string
          title: Order No
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductionInfoResponse'
                title: Response Get Production Info By Order No Api Production  Order
                  No  Get
              example:
              - id: 1
                order_no: ORD20250101001
                model: MODEL-A
                brand_no: BRAND-001
                quantity: 100
                job_type: 组装
                worklog_no: WL001
                performance_factor: 1.2
                upload_date: '2025-01-01T00:00:00Z'
                created_at: '2025-01-01T00:00:00Z'
                updated_at: '2025-01-01T00:00:00Z'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '404':
          description: 订单号不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AuthenticationErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: AuthenticationErrorResponse
      description: 认证错误响应模型
      example:
        detail: {}
        error: true
        message: 认证失败
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InternalServerErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: InternalServerErrorResponse
      description: 服务器内部错误响应模型
      example:
        detail:
          message: Connection timeout
          type: DatabaseError
        error: true
        message: 服务器内部错误
    NotFoundErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: NotFoundErrorResponse
      description: 资源未找到错误响应模型
      example:
        detail:
          id: '12345'
          resource: order
        error: true
        message: 资源未找到
    ProductionInfoListResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        items:
          items:
            $ref: '#/components/schemas/ProductionInfoResponse'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - page_size
      - items
      title: ProductionInfoListResponse
      description: 生产信息列表响应模型
    ProductionInfoResponse:
      properties:
        id:
          type: integer
          title: Id
          description: 记录 ID
        order_no:
          type: string
          title: Order No
          description: 订单号
        model:
          type: string
          title: Model
          description: 型号
        brand_no:
          type: string
          title: Brand No
          description: 品牌编号
        quantity:
          type: integer
          title: Quantity
          description: 数量
        job_type:
          type: string
          title: Job Type
          description: 作业类型
        worklog_no:
          type: string
          title: Worklog No
          description: 工作量编号
        performance_factor:
          type: number
          title: Performance Factor
          description: 绩效系数
        upload_date:
          type: string
          format: date-time
          title: Upload Date
          description: 上传日期
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: 更新时间
      type: object
      required:
      - id
      - order_no
      - model
      - brand_no
      - quantity
      - job_type
      - worklog_no
      - performance_factor
      - upload_date
      - created_at
      - updated_at
      title: ProductionInfoResponse
      description: 生产信息响应模型
      example:
        brand_no: BRAND-001
        created_at: '2025-01-01T00:00:00Z'
        id: 1
        job_type: 组装
        model: MODEL-A
        order_no: ORD20250101001
        performance_factor: 1.2
        quantity: 100
        updated_at: '2025-01-01T00:00:00Z'
        upload_date: '2025-01-01T00:00:00Z'
        worklog_no: WL001
...
    User:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 1
          title: Username
        id:
          type: integer
          title: Id
          description: 用户 ID
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: 更新时间
      type: object
      required:
      - username
      - id
      - created_at
      - updated_at
      title: User
      description: 用户响应模型（不包含敏感信息）
      example:
        created_at: '2025-01-01T00:00:00Z'
        id: 1
        updated_at: '2025-01-01T00:00:00Z'
        username: admin
    UserLogin:
      properties:
        username:
          type: string
          maxLength: 50
          minLength: 1
          title: Username
          description: 用户名
        password:
          type: string
          minLength: 6
          title: Password
          description: 密码
      type: object
      required:
      - username
      - password
      title: UserLogin
      description: 用户登录模型
      example:
        password: password123
        username: admin
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValidationErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: ValidationErrorResponse
      description: 验证错误响应模型
      example:
        detail:
          field: order_no
          reason: 订单号不能为空
        error: true
        message: 数据验证失败
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入 JWT token，格式：Bearer <token>
...